# rate-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for notifying exchange rate",
        "title": "Exchange rate notifier API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/subscribe": {
            "post": {
                "description": "subscribe to notifications",
                "tags": [
                    "subscription"
                ],
                "summary": "Subscribe to notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "invalid email format",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "409": {
                        "description": "subscription already exists",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "failed to create subscription",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/unsubscribe": {
            "post": {
                "description": "unsubscribe from notifications",
                "tags": [
                    "subscription"
                ],
                "summary": "Unsubscribe from notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "invalid email format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to delete subscription",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/rate": {
            "get": {
                "description": "Get the current exchange rate from the external API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rate"
                ],
                "summary": "Get current exchange rate",
                "responses": {
                    "200": {
                        "description": "Returns the current exchange rate",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to get exchange rate",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}